+++ apps/base/src/components/slots/landingPage/main.tsx
@@ -1,6 +1,0 @@
-// Slot: UI/landingPage
-import BasicLandingPage from "@ulld/landing-layouts/beta"
-
-
-
-export default BasicLandingPage