+++ packages/api/src/types.ts
@@ -1,0 +1,10 @@
+import { MdxNoteSummaryOutputWithMdxTransforms } from "./classes/prismaMdxRelations/schemas/withMdxTransforms";
+
+export interface SearchResultsPageProps {
+    notes: MdxNoteSummaryOutputWithMdxTransforms[];
+}
+
+
+export type BookmarksResultsPageProps = Pick<SearchResultsPageProps, "notes">
+
+export type { MdxNoteSummaryOutputWithMdxTransforms };