+++ packages/bib-manager/src/types.ts
@@ -1,0 +1,5 @@
+import {serverClient} from "@ulld/api/serverClient"
+
+export interface BibliographyPageProps {
+   prismaBib: Awaited<ReturnType<typeof serverClient.bibliography.getBib>>
+}