+++ packages/configschema/src/__scripts__/generateSlotMap.ts
@@ -2,8 +2,20 @@ import { globSync } from "glob";
 import path from "path";
 import fs from "fs";
 import { capitalize } from "@ulld/utilities/stringUtils";
+import { slotTypes } from "@ulld/utilities/types";
+import propsExtendsMap from "@ulld/utilities/buildStaticData"
+import { gatherProtectedPaths } from "./gatherProtectedPaths";
 
-const targetDir = path.join(__dirname, "../../../../apps/base/src");
+const testRoot = process.env.ULLD_TEST_ROOT;
+
+if (!testRoot) {
+    throw new Error(
+        "ULLD_TEST_ROOT environment variable ws not present at run time in the generateSlotMap script.",
+    );
+}
+
+const targetDir = path.join(testRoot, "/src");
+
 const mdxPath = path.join(
     __dirname,
     "../../../../apps/website/content/embeddedDocs/developer/slotMap.txt",
@@ -23,26 +35,47 @@ let items: {
     subSlot: string;
     clientOnly: boolean;
     inReduxProvider: boolean;
+    type?: string;
+    propsExtends?: string;
 }[] = [];
+
+gatherProtectedPaths(files.map((f) => path.join(targetDir, f)))
 
 for (const k of files) {
     let content = fs.readFileSync(path.join(targetDir, k), { encoding: "utf-8" });
     let lineOne = content.split("\n")[0].trim();
     if (lineOne.startsWith("// Slot:")) {
-        let slots = lineOne
-            .replace("// Slot: ", "")
-            .trim()
+        let words = lineOne.replace("// Slot: ", "").split(" ");
+        let slots = words
+            .find((w) => w.includes("/"))
             .split("/")
-            .map((f) => f.trim());
+            .map((s) => s.trim());
+        let type = words
+            .find((w) => w.startsWith("type:"))
+            ?.replace("type:", "")
+            .trim() || "component"
+        let propsExtends = words
+            .find((w) => w.startsWith("propsExtends:"))
+            ?.replace("propsExtends:", "")
+            .trim();
         if (slots.length !== 2) {
             throw new Error(`Misformed slot string at ${k}`);
+        }
+        if(!slotTypes.includes(type as any)){
+            throw new Error(`No slotData type found for ${type}.`)
+        }
+        if(Boolean(propsExtends && !(propsExtends in propsExtendsMap.propsExtendsMap))){
+            throw new Error(`Slot prop can not extend ${propsExtends}. It was is not included in the propsExtendsMap object.`)
         }
         const newItem = {
             path: k,
             parentSlot: slots[0],
             subSlot: slots[1],
-            clientOnly: content.includes("Slot:clientOnly"),
-            inReduxProvider: content.includes("Slot:inReduxProvider"),
+            clientOnly: content.includes("clientOnly"),
+            inReduxProvider: content.includes("inReduxProvider"),
+            relativePath: `src/${k}`,
+            type,
+            propsExtends,
         };
         let hasItem = items.find((f) =>
             Boolean(
@@ -69,7 +102,10 @@ let slotMap: Record<
     string,
     Record<
         string,
-        Pick<(typeof items)[0], "inReduxProvider" | "clientOnly" | "path">
+        Pick<
+            (typeof items)[0],
+            "inReduxProvider" | "clientOnly" | "path" | "type" | "propsExtends"
+        >
     >
 > = {};
 
@@ -78,9 +114,11 @@ for (const k of items) {
         slotMap[k.parentSlot] = {};
     }
     slotMap[k.parentSlot][k.subSlot] = {
-        path: k.path,
+        path: `src/${k.path}`,
         clientOnly: k.clientOnly,
         inReduxProvider: k.inReduxProvider,
+        type: k.type,
+        propsExtends: k.propsExtends,
     };
 }
 
@@ -89,20 +127,6 @@ const targetPath = path.join(
     "../../../utilities/src/utils/slotMap.json",
 );
 
-const wrapMdxContent = (val: string) => {
-    const d = new Date();
-    return `---
-title: SlotMap Type
-id: slotMapType
-created: 6-24-24
-updated: ${d.getMonth()}-${d.getDate()}-${d.getFullYear()}
----
-
-\`\`\`tsx
-${val}
-\`\`\`
-`;
-};
 
 let mdxSlotMap = `type SlotMap = {
 `;
@@ -159,11 +183,14 @@ ${Object.values(slotSubKeys).join("\n\n")}
 
 `;
 
-
 fs.writeFileSync(mdxPath, mdxSlotMap, { encoding: "utf-8" });
 
-fs.writeFileSync(typeRootPath, `export ${mdxSlotMap}
-`, {encoding: "utf-8"})
+fs.writeFileSync(
+    typeRootPath,
+    `export ${mdxSlotMap}
+`,
+    { encoding: "utf-8" },
+);
 
 fs.writeFileSync(
     typePath,
@@ -173,7 +200,9 @@ export type SlotMap = SM
 
 export type PluginSlotKey = keyof SM
 
-${Object.keys(slotSubKeys).map((k) => `export type ${capitalize(k)}SubSlots = keyof SM["${k}"]`).join("\n\n")}`,
+${Object.keys(slotSubKeys)
+        .map((k) => `export type ${capitalize(k)}SubSlots = keyof SM["${k}"]`)
+        .join("\n\n")}`,
     { encoding: "utf-8" },
 );