+++ packages/configschema/src/developer/pluginPageSchema.ts
@@ -1,9 +1,8 @@
 import {z} from 'zod'
 import { transformExportString } from './transforms'
-import { PluginPageConfig } from './types/pageConfig'
 
 
-export const pluginAdditionalPageSchema: z.ZodType<PluginPageConfig> = z.object({
+export const pluginAdditionalPageSchema = z.object({
     targetUrl: z.string().describe("The target URL to place this page at. This is synonomous with a file path from the root of the app directory, including intercepted routes. An intercepted modal route for example should appear as `@modal/(.)myPath/...` even though `@modal` doesn't appear in the URL. This must be unique, as if it overwrites an existing route it will not be applied.").transform((f) => {
         let _f = f
         if(_f.endsWith("page.tsx")){
@@ -17,6 +16,7 @@ export const pluginAdditionalPageSchema: z.ZodType<PluginPageConfig> = z.object(
         }
         return _f
     }),
+    slot: z.string().optional().describe("The optional slot key that matches a corresponding slot of type page."),
     export: z.string().describe("The export in your package.json that matches a single component that will act as this page. The component must be the default export, and will be passed all props that the page receives, like params and searchParams. This can include intercepted modal routes with the @modal/(.)... syntax, nested routes and parameter based routes with the [someParam] syntax.").transform(transformExportString),
     exportsPageProps: z.boolean().default(false).describe("Whether or not the file at the export field exports a type named PageProps to apply to the parent page properties. This type should include a params property and/or a searchParams property as they apply to the page itself. This is mostly for the sake of completeness, but can help other developers to work in a bug free environment if they choose to extend their own app."),
 })