+++ packages/task-manager/src/types.ts
@@ -1,0 +1,10 @@
+import { UniqueTaggables } from "@ulld/utilities/types";
+
+export interface TaskDetailsPageProps {
+    isModal: boolean;
+}
+
+export interface AddTaskProps {
+    isModal: boolean;
+    existingTaggables: UniqueTaggables;
+}