+++ packages/utilities/src/types/corePageUtils/commonPagePropTypes.ts
@@ -1,0 +1,5 @@
+export type UniqueTaggables = {
+    tags: string[];
+    subjects: string[];
+    topics: string[];
+}