+++ packages/utilities/src/types/slotDataType.ts
@@ -1,0 +1,16 @@
+import propsExtends from "../utils/buildStaticData.json"
+
+export const slotTypes = [
+    "page",
+    "component",
+    "modalPage"
+] as const
+
+
+export interface SlotDataType {
+    path: string
+    clientOnly?: boolean
+    inReduxProvider?: boolean
+    type?: typeof slotTypes[number]
+    propsExtends?: keyof typeof propsExtends["propsExtendsMap"]
+}