+++ packages/database/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ulld/database
 
+## 0.0.6
+
+### Patch Changes
+
+- Removed all unpublished packages dynamically.
+- Updated dependencies
+  - @ulld/typescript-config@0.0.5
+
+## 0.0.5
+
+### Patch Changes
+
+- One more time with this cluster\*\*\*\* of dependencies.
+- Updated dependencies
+  - @ulld/typescript-config@0.0.4
+
 ## 0.0.4
 
 ### Patch Changes