+++ packages/developer-cli/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ulld/developer-cli
 
+## 0.0.6
+
+### Patch Changes
+
+- Removed all unpublished packages dynamically.
+- Updated dependencies
+  - @ulld/developer-schemas@0.0.6
+  - @ulld/developer@0.0.6
+
+## 0.0.5
+
+### Patch Changes
+
+- One more time with this cluster\*\*\*\* of dependencies.
+- Updated dependencies
+  - @ulld/developer-schemas@0.0.5
+  - @ulld/developer@0.0.5
+
 ## 0.0.4
 
 ### Patch Changes