+++ packages/developer-schemas/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ulld/developer-schemas
 
+## 0.0.6
+
+### Patch Changes
+
+- Removed all unpublished packages dynamically.
+
+## 0.0.5
+
+### Patch Changes
+
+- One more time with this cluster\*\*\*\* of dependencies.
+
 ## 0.0.4
 
 ### Patch Changes