+++ packages/developer/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @ulld/developer
 
+## 0.0.6
+
+### Patch Changes
+
+- Removed all unpublished packages dynamically.
+- Updated dependencies
+  - @ulld/typescript-config@0.0.5
+  - @ulld/eslint-config@0.0.6
+  - @ulld/configschema@0.0.6
+  - @ulld/database@0.0.6
+  - @ulld/config@0.0.6
+  - @ulld/logger@0.0.6
+
+## 0.0.5
+
+### Patch Changes
+
+- One more time with this cluster\*\*\*\* of dependencies.
+- Updated dependencies
+  - @ulld/typescript-config@0.0.4
+  - @ulld/eslint-config@0.0.5
+  - @ulld/configschema@0.0.5
+  - @ulld/database@0.0.5
+  - @ulld/config@0.0.5
+  - @ulld/logger@0.0.5
+
 ## 0.0.4
 
 ### Patch Changes