+++ packages/embeddable-components/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @ulld/embeddable-components
 
+## 0.0.6
+
+### Patch Changes
+
+- Removed all unpublished packages dynamically.
+- Updated dependencies
+  - @ulld/default-image-map@0.0.6
+  - @ulld/component-configs@0.0.6
+  - @ulld/tailwind@0.0.6
+  - @ulld/configschema@0.0.6
+  - @ulld/utilities@0.0.6
+  - @ulld/render@0.0.6
+  - @ulld/hooks@0.0.6
+  - @ulld/icons@0.0.6
+  - @ulld/state@0.0.6
+  - @ulld/api@0.0.6
+
+## 0.0.5
+
+### Patch Changes
+
+- One more time with this cluster\*\*\*\* of dependencies.
+- Updated dependencies
+  - @ulld/default-image-map@0.0.5
+  - @ulld/component-configs@0.0.5
+  - @ulld/tailwind@0.0.5
+  - @ulld/configschema@0.0.5
+  - @ulld/utilities@0.0.5
+  - @ulld/render@0.0.5
+  - @ulld/hooks@0.0.5
+  - @ulld/icons@0.0.5
+  - @ulld/state@0.0.5
+  - @ulld/api@0.0.5
+
 ## 0.0.4
 
 ### Patch Changes