+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @ulld/hooks
 
+## 0.0.6
+
+### Patch Changes
+
+- Removed all unpublished packages dynamically.
+- Updated dependencies
+  - @ulld/component-map@0.0.6
+  - @ulld/parsers@0.0.6
+  - @ulld/api@0.0.6
+
+## 0.0.5
+
+### Patch Changes
+
+- One more time with this cluster\*\*\*\* of dependencies.
+- Updated dependencies
+  - @ulld/component-map@0.0.5
+  - @ulld/parsers@0.0.5
+  - @ulld/api@0.0.5
+
 ## 0.0.4
 
 ### Patch Changes