+++ packages/icons/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @ulld/icons
 
+## 0.0.6
+
+### Patch Changes
+
+- Removed all unpublished packages dynamically.
+- Updated dependencies
+  - @ulld/tailwind@0.0.6
+  - @ulld/utilities@0.0.6
+  - @ulld/hooks@0.0.6
+
+## 0.0.5
+
+### Patch Changes
+
+- One more time with this cluster\*\*\*\* of dependencies.
+- Updated dependencies
+  - @ulld/tailwind@0.0.5
+  - @ulld/utilities@0.0.5
+  - @ulld/hooks@0.0.5
+
 ## 0.0.4
 
 ### Patch Changes